==> notebooks/1_Data_Fetching.ipynb <==
# 1. Data Fetching
In this section we will fetch a dataset from Wikidata where the following experiments will take place. Our experimental dataset will consist of a set of the most important entities from the most important classes of Wikidata. To calculate the importance of each entity we will use their PageRank value, and the importance of each class will be calculated using [ClassRank](https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0252862).

The main steps of this data fetching process will be as follows:
1. We will start with a list of pagerank scores from Wikidata
2. Retrieve all the classes of each entity included in the pagerank scores list.
3. Calculate the classrank score of each class.
4. Get the most important classes and instances of each class based on their classrank and pagerank scores, respectively.
5. Fetch the edit history of those instances (see scripts folder).

## Notebook setup
### Importing libraries
We will start by importing the libraries used in the notebook. These are some of the external libraries required to run the notebook:
- __SPARQLWrapper__ will be used to execute SPARQL queries to fetch information about the entities from Wikidata.
- __lightrdf__ will be used to load the information fetched from Wikidata in an efficient way.
- __plotly__ will be the main plotting library used.

from requests.adapters import HTTPAdapter
from SPARQLWrapper import JSON, SPARQLWrapper, RDFXML, POST
from urllib3.util.retry import Retry

import lightrdf
import rdflib
from rdflib.compare import graph_diff

import json
import math
import os
import pickle
import requests
import sys
import time

sys.path.insert(0, "classrank")

from helpers.classrank import generate_classrank

### Definition of constants
We will also define the following constants that will be used along the notebook:

DATA_DIR = os.path.join('..', 'data')
OUTPUT_DIR = os.path.join('output', '1_data_fetching')

CLASSES_FILE = os.path.join(OUTPUT_DIR, 'top_classes.pkl')
CLASSRANK_FILE = os.path.join(OUTPUT_DIR, 'classrank.json')
ENTITIES_FILE = os.path.join(OUTPUT_DIR, 'entities_ids.txt')
ENTITIES_CLASSES_FILE = os.path.join(OUTPUT_DIR, 'entities_classes_ids.csv')
GRAPH_FILE = os.path.join(OUTPUT_DIR, 'entities_file.ttl')

PAGERANK_SCORES_DIR = os.path.join(DATA_DIR, 'pagerank')
PAGERANK_SCORES_FILE = os.path.join(PAGERANK_SCORES_DIR, '2020-11-14.allwiki.links.rank')

CLASS_POINTER = 'wdt:P31'

WIKIDATA_API_ENDPOINT = "http://www.wikidata.org/w/api.php"
WIKIDATA_BASE = "http://www.wikidata.org/entity/"

### Defining functions and classes

The following classes will be used to represent the information fetched in the notebook:

from dataclasses import dataclass
from typing import List

@dataclass
class KGEntity:
    qid: str
    pagerank_score: float

@dataclass
class KGClass:
    name: str
    qid: str
    classrank_score: float
    instances: List[KGEntity]


The following functions will be used to execute SPARQL queries:

def build_name_query(entities):
    return f"""
    SELECT ?item ?itemLabel
    WHERE {{
        VALUES ?item {{ {' '.join(entities)} }}
        SERVICE wikibase:label {{ bd:serviceParam wikibase:language "[AUTO_LANGUAGE],en". }}
    }}
    """

def build_instanceof_query(entities):
    return f"""
    CONSTRUCT {{
        ?subj wdt:P31 ?obj
    }}
    WHERE {{
        VALUES ?subj {{ {' '.join(entities)} }}
        ?subj wdt:P31 ?obj .
    }}
    """

def execute_query(sparql_query, return_format=RDFXML):
    sparql = SPARQLWrapper("https://query.wikidata.org/sparql")

    sparql.setMethod(POST)  
    sparql.setQuery(sparql_query)
    sparql.setReturnFormat(return_format)

    # exponential backoff
    retries = 0
    while True:
        try:
            return sparql.query().convert()
        except:
            sleep = 2 ** retries
            print(f"Backing off for {sleep} seconds...")
            time.sleep(sleep)
            retries += 1


## Fetching PageRank scores

with open(PAGERANK_SCORES_FILE, 'r', encoding='utf-8') as f:
    pagerank_scores = {"wd:{}".format(line.split('\t')[0]): float(line.split('\t')[1])
                       for line in f.readlines()}

print("Number of entities with a pagerank score: ", len(pagerank_scores))

## Obtaining a subset from Wikidata

batch_size = 50000

entities = list(pagerank_scores.keys())
num_batches = math.ceil(len(entities) / batch_size)

with open(GRAPH_FILE, 'wb') as f:
    for i in range(num_batches):
        time_start = time.time()

        start, end = (i * batch_size, (i+1) * batch_size)
        print(f"Batch: {start} to {end}")
        batch_entities = entities[i*batch_size:(i+1)*batch_size]
        query = build_instanceof_query(batch_entities)
        graph = execute_query(query)
        f.write(graph.serialize(format='turtle'))
        
        time_end = time.time()
        print(f"{time_end - time_start} s\n")

## Getting the most important classes

### Calculating ClassRank
In the following cells we are going to use the pagerank values that we have loaded before to compute the classrank values of each class.

__WARNING:__ The following cells make take a bit of time to compute (~3 min). Feel free to skip to the following section where we load the precomputed classrank values if you want to skip these computations.

parser = lightrdf.Parser()

subjects = set([str(t[0]) for t in parser.parse(GRAPH_FILE, base_iri=None)])

num_entities_pagerank = len(pagerank_scores)
num_entities_queried = len(subjects)

print("Original number of entities: ", num_entities_pagerank)
print("Final number of entities (with class information): ", num_entities_queried)

pagerank_scores = {k: v for k, v in pagerank_scores.items()
                   if f"wd:{k[3:]}" in subjects}

del subjects

classrank_str = generate_classrank(graph_file=GRAPH_FILE, pagerank_scores=pagerank_scores,
    raw_classpointers=CLASS_POINTER,
    save_memory_mode=True, string_return=True)

classrank = json.loads(classrank_str)

with open(CLASSRANK_FILE, 'w', encoding='utf-8') as f:
    json.dump(classrank, f)

### Loading precomputed ClassRank values

with open(CLASSRANK_FILE, 'r', encoding='utf-8') as f:
    classrank = json.load(f)

### Getting the most important classes based on ClassRank

top_classes_cr = sorted(classrank, key=lambda x: x['CR_score'], reverse=True)[:NUM_CLASSES]

print(f"Most important class: {classrank[0]['class']} - score={classrank[0]['CR_score']}")

del classrank

top_classes = []
batch_size = 10000

for c in top_classes_cr:
    print("Fetching data for class: ", c['class'])
    
    name_query = build_name_query([c['class']])
    class_name_json = execute_query(name_query, return_format=JSON)
    class_name = class_name_json['results']['bindings'][0]['itemLabel']['value']

    if 'wiki' in class_name.lower():
        print("Skipping class: ", class_name)
        continue
    
    # retrieve data from top class instances
    all_class_instances = c['cps'][CLASS_POINTER]
    
    entities_pagerank = [(e, pagerank_scores[e]) for e in all_class_instances]

    """
    num_batches = math.ceil(len(entities_pagerank) / batch_size)
    entities_names = []
    for i in range(num_batches):
        start, end = (i * batch_size, (i+1) * batch_size)
        name_query = build_name_query([e[0] for e in entities_pagerank[start:end]])
        entities_names_json = execute_query(name_query, return_format=JSON)
        entities_names += [e['itemLabel']['value'] for e in entities_names_json['results']['bindings']]
    """
    class_kgentities = [KGEntity(qid=item[0].replace('wd:', ''), pagerank_score=item[1])
                        for item in entities_pagerank]
    
    kgclass = KGClass(name=class_name, qid=c['class'].replace('wd:', ''),
                      classrank_score=c['CR_score'], instances=class_kgentities)
    top_classes.append(kgclass)

## Saving results

with open(CLASSES_FILE, 'rb') as f:
    top_classes = pickle.load(f)

import csv
    
with open(ENTITIES_CLASSES_FILE, 'w') as csvfile:
    fieldnames = ['entity_id', 'class_id']
    writer = csv.DictWriter(csvfile, fieldnames=fieldnames)
    writer.writeheader()
    for kg_class in top_classes:
        class_qid = kg_class.qid
        for entity in kg_class.instances:
            writer.writerow({'entity_id': entity.qid, 'class_id': class_qid})

with open(CLASSES_FILE, 'wb') as f:
    pickle.dump(top_classes, f)

with open(ENTITIES_FILE, 'w') as f:
    for c in top_classes:
        all_class_instances_ids = [i.qid for i in c.top_entities]
        f.write('\n'.join(all_class_instances_ids))
