==> notebooks/2_Data_Exploration.ipynb <==
# Data exploration
In this notebook we will analyse the revisions data that has been indexed in the MongoDB instance. This index process was carried out with the programs from the 'scripts' folder. We will create datasets with general information and statistics from this data that will be used to both understand the edits behaviours in Wikidata and also to create predictive models from these datasets later on. 

## Notebook setup
We will start by preparing the notebook. The following cells must be executed if there is a problem saving the figures that are created in the notebook:

!pip install -U kaleido

!jupyter lab build

### Imports
Now we will import the general modules used accross the notebook:

import json
import math
import os
import pickle
import requests

from collections import defaultdict

import numpy as np
import pandas as pd
import plotly.express as px
import plotly.graph_objects as go
import plotly.io as pio

from tqdm.auto import tqdm

op_types = ['add', 'remove', 'replace']
op_colors = {'add': px.colors.qualitative.Plotly[0], 'remove': px.colors.qualitative.Plotly[1], 'replace': px.colors.qualitative.Plotly[2]}
pio.templates.default = "plotly_white"

### Constants
These will be the constants that are reused throughout the notebook. They mainly consist of directories where data will be read from or written into:

DATA_DIR = os.path.join('..', 'data')

NOTEBOOK1_OUTPUT_DIR = os.path.join('output', '1_data_fetching')
CLASSES_FILE = os.path.join(NOTEBOOK1_OUTPUT_DIR, 'top_classes.pkl')

OUTPUT_DIR = os.path.join('output', '2_data_exploration')
CHARTS_OUTPUT_DIR = os.path.join(OUTPUT_DIR, 'charts')
CSV_OUTPUT_DIR = os.path.join(OUTPUT_DIR, 'csv')
DATAFRAMES_OUTPUT_DIR = os.path.join(OUTPUT_DIR, 'df_results')

WIKIDATA_BASE = "https://www.wikidata.org/w/api.php"

### Auxiliary functions
To finalize the setup of this notebook, we will define a set of common functions. The first set of functions will be used to fetch labels of property ids from Wikidata. The second set will be used to serialize and deserialize data with pickle: 

props_name_cache = {}


def update_props_name_cache(props_pids):
    props = '|'.join(props_pids)
    url = f'{WIKIDATA_BASE}?action=wbgetentities&format=json&props=labels&ids={props}&languages=en'
    r = requests.get(url)
    data = r.json()
    for k, v in data['entities'].items():
        if 'missing' in v:
            props_name_cache[k] = f"[deleted property] ({v['id']})"
        else:
            try:
                props_name_cache[k] = v['labels']['en']['value']
            except KeyError:
                props_name_cache[k] = f"[No english label] ({v['id']})"

def get_name_of_props(props_pids):
    props_to_query = [p for p in props_pids if p not in props_name_cache]
    if len(props_to_query) > 0:
        update_props_name_cache(props_to_query)
    return [props_name_cache[p] for p in props_pids]


def load_df(filename):
    with open(os.path.join(DATAFRAMES_OUTPUT_DIR, f"{filename}.pkl"), 'rb') as f:
        return pickle.load(f)

def save_df(df, filename):
    with open(os.path.join(DATAFRAMES_OUTPUT_DIR, f"{filename}.pkl"), 'wb') as f:
        pickle.dump(df, f)

def save_csv(df, filename):
    with open(os.path.join(CSV_OUTPUT_DIR, f"{filename}.csv"), 'w') as f:
        df.to_csv(f, index=False)

def save_fig(fig, filename):
    fig.write_html(os.path.join(CHARTS_OUTPUT_DIR, f"{filename}.html"), include_plotlyjs=False)
    fig.write_image(os.path.join(CHARTS_OUTPUT_DIR, f"{filename}.svg"))
    fig.write_image(os.path.join(CHARTS_OUTPUT_DIR, f"{filename}.pdf"))


## Loading the data from the previous notebook
To begin with this notebook we are going to load the classes that compose our dataset, which were serialized in the previous notebook:

from dataclasses import dataclass
from typing import List

@dataclass
class KGEntity:
    qid: str
    pagerank_score: float

@dataclass
class KGClass:
    name: str
    qid: str
    classrank_score: float
    instances: List[KGEntity]


with open(CLASSES_FILE, 'rb') as f:
    top_classes = pickle.load(f)

print(sum([len(c.instances) for c in top_classes]))

We can see how our dataset is composed of 9693840 entities.

## Displaying general information

We are going to perform now a quick exploratory analysis to confirm that the data has been loaded correctly. After executing the next cell we can see which are the top 25 classes of our dataset based on their classrank score:

print("Top 25 classes based on classrank")
print("-" * 35)
print('\n'.join([f"{c.name} - score: {c.classrank_score}" for c in top_classes[:25]]))

We are now going to plot the distribution of instances accross each class:

# Histogram
x = [len(c.instances) for c in top_classes]
labels = {
    "x": "Number of class instances"
}

fig = px.histogram(x=x, template='plotly_white', labels=labels)
fig.update_layout(bargap=0.02)
fig.show()

We can see from the figure above that the mayority of classes have less than 200K instances, while there are some isolated classes with more instances, with a class having 3.8 millions of instances. We will finish this initial exploration saving the figure:

save_fig(fig, '0_histogram')

## Querying the diff data
We are now going to query the revision data that has been indexed to the database. Our database will contain the following two collections:
- __wd_entities__: This collection contains the current entity data of our dataset. It represents the data model of Wikidata, with each entity having the following schema:
    - id: Internal Wikidata identificator of the entity (e.g. 2491849209).
    - entity_id: 'Public' id of the entity (e.g. Q42).
    - class_ids: List of ids of the classes the entity belongs to (e.g. \[Q5, Q20\]).
    - entity_json: Complete JSON content of the entity in Wikidata (see [this link](https://doc.wikimedia.org/Wikibase/master/php/md_docs_topics_json.html) for more information about the JSON schema followed).
- __wd_revisions__: This collection contains information about all the revisions made to each entity from the dataset. It follows this schema:
    - id: Internal Wikidata identificator of the revision (e.g. 129529015819).
    - entity_id: 'Public' id of the entity this revision refers tol (e.g. Q42).
    - class_ids: List of ids of the classes the entity belongs to (e.g. \[Q5, Q20\]).
    - parent_id: Internal Wikidata identificator of the previous revision made to the entity (e.g. 31980681969).
    - timestamp: Timestamp of the revision, following the 'YYYY-MM-DDThh:mm:ssZ' format (e.g. 2014-08-28T15:45:20Z).
    - username: Username of the user that made the revision (e.g. 'Axipbot').
    - comment: Comment of the user that made the revision, if any.
    - entity_diff: List of operations made to the Wikidata entity in the revision, following the [JSON Patch](http://jsonpatch.com) format.

An example value of the entity diff array can be seen here:
```json
{
    "op": "replace",
    "path": "/claims/P150/7/mainsnak/datavalue/value/numeric-id",
    "value": 31326
}
```

### Connecting to the database
We will now make a connection to the Mongo database where this data is stored. If everything was set up from the docker-compose file, this connection will be configured automatically:

import getpass
import os
import pprint


def env_or_callback(env_name, callback, *args):
    return os.getenv(env_name) if os.getenv(env_name) is not None else callback(*args)

from pymongo import MongoClient
import pymongo

MONGO_USERNAME = env_or_callback("MONGO_USERNAME", input, "Mongo username: ")
MONGO_PASSWORD = env_or_callback("MONGO_PASSWORD", getpass.getpass, "Mongo password: ")
MONGO_URL = env_or_callback("MONGO_URL", input, "Mongo url: ")
MONGO_DATABASE = "wd_diff"

def get_database():
    # Provide the mongodb atlas url to connect python to mongodb using pymongo
    CONNECTION_STRING = f"mongodb://{MONGO_USERNAME}:{MONGO_PASSWORD}@127.0.0.1:27017/{MONGO_DATABASE}"

    # Create a connection using MongoClient. You can import MongoClient or use pymongo.MongoClient
    client = MongoClient(CONNECTION_STRING)

    # Create the database for our example (we will use the same database throughout the tutorial
    return client[MONGO_DATABASE]

db = get_database()
wd_entities = db.wd_entities
wd_revisions = db.wd_revisions

### Indexes creation
Before querying the data we are going to create indexes that support our queries to speed up their execution. We will be mainly querying information based on their entity id or their class id, so we will be making indexes on those fields. This step may take a while depending on the i/o speed of the disk where Mongo data is stored.

If the indexes where already created, this cell will not modify anything.

wd_revisions.create_index([("class_ids", pymongo.DESCENDING)])
wd_revisions.create_index([("entity_id", pymongo.DESCENDING)])
wd_entities.create_index([("class_ids", pymongo.DESCENDING)])
wd_entities.create_index([("entity_id", pymongo.DESCENDING)])

We will now list the indexes of both collections to check if they have been created correctly:

list(wd_entities.list_indexes())

list(wd_revisions.list_indexes())

### A. Descriptive analysis
With our connection to the database ready we can start exploring the data. The steps of each analysis will always be very similar:
1. Define a set of functions used to query the information we need.
2. Execute the queries and store the information in a Dataframe.
3. Briefly explore and describe the Dataframe.
4. Visualize the contents of the Dataframe.

Since in most cases the **step number 2 may take a long time to execute, we provide the precomputed results** to avoid these execution times and continue the execution of the notebook. These precomputed results are usually provided in the cell following to the long computation one.

To begin we will perform a descriptive analysis of the *operations* that compose our dataset. It is important to first mention the differences between *operation* and *revision* in our context. A revision is analogous to the one in Wikidata, where a user modifies the contents of an entity. One revision can be composed of many *operations* which can be seen as atomic changes to an entity. An operation could belong to one of three different types: 'add', 'remove' or 'replace'.

The following cell defines two functions that return the number of operations of instances of a class in our data. The first one aggregates all the operations, while the second one performs the aggregation based on each operation type:

def get_number_of_operations_per_instance(collection, class_id):
    pipeline = [
        {"$match": {"class_ids": class_id}},
        {"$unwind": "$entity_diff"},
        {
            "$group":
            {
                "_id": "$entity_id",
                "count": {"$sum": 1}
            }
        }
    ]
    return list(collection.aggregate(pipeline, allowDiskUse=True))

def get_number_of_op_types_per_instance(collection, class_id):
    pipeline = [
        {"$match": {"class_ids": class_id}},
        {"$unwind": "$entity_diff"},
        {
            "$group":
            {
                "_id": {"entity": "$entity_id", "op": "$entity_diff.op"},
                "count": {"$sum": 1}
            }
        }
    ]
    return list(collection.aggregate(pipeline, allowDiskUse=True))

With the functions ready we will now create our first Dataframe with statistics about the operations of each class.

> ⚠️⏰ This cell may take a long time of execute (est. 2h, can vary greatly based on disk speed), the precomputed result is available in the following cell and can be directly loaded without executing this cell.

data = {
    'class': [],
    'op': [],
    'mean': [],
    'median': [],
    'std': [],
    'max': [],
    'min': []
}

ops = ['add', 'remove', 'replace']

with tqdm(top_classes) as t:
    for kg_class in t:
        t.set_description(f"Class: {kg_class.name}")
        num_ops_instances = get_number_of_op_types_per_instance(wd_revisions, kg_class.qid)
        for op in ops:
            op_counts = [d['count'] for d in num_ops_instances if d['_id']['op'] == op]
            data['mean'].append(np.mean(op_counts))
            data['max'].append(np.max(op_counts))
            data['min'].append(np.min(op_counts))
            data['std'].append(np.std(op_counts))
            data['median'].append(np.median(op_counts))
            data['class'].append(kg_class.name)
            data['op'].append(op)

descriptives_df = pd.DataFrame(data)
save_df(descriptives_df, 'a_descriptives')

Execute the following cell to load the precomputed dataframe:

descriptives_df = load_df('a_descriptives')

Now that the dataframe has been loaded, we will quickly check its contents:

descriptives_df.head()

### B. Number of documents
In this section we will briefly explore the number of documents in our dataset. The following functions return respectively the total number of documents of a collection and the number of documents of a collection of a given class.

def get_doc_count(collection):
    return collection.estimated_document_count()

def get_num_instances_of_class(collection, class_id):
    return collection.count_documents({"class_ids": class_id})

We can now execute the functions to check the number of entities and revisions in our dataset:

print(f"Number of entities: {get_doc_count(wd_entities)}")
print(f"Number of revisions: {get_doc_count(wd_revisions)}")

In the following cell we are going to fetch the number of instances of each class indexed in our database, so we can compare them to the total number of instances of that class.

> This cell should execute quickly (<1min), but we also provide the precomputed dataset in the following cell in case it's necessary.

data = {
    'class': [],
    'num_instances': [],
    'indexed_instances': []
}

with tqdm(top_classes) as t:
    for kg_class in t:
        t.set_description(f"Class: {kg_class.name}")
        indexed_class_instances = get_num_instances_of_class(wd_entities, kg_class.qid)
        data['class'].append(kg_class.name)
        data['num_instances'].append(len(kg_class.instances))
        data['indexed_instances'].append(indexed_class_instances)

num_instances_df = pd.DataFrame(data)
save_df(num_instances_df, 'b_instances')

num_instances_df = load_df('b_instances')

from IPython.display import HTML

HTML(num_instances_df.sort_values(by='num_instances', ascending=False).head(n=10).to_html(index=False))

If we see the table above we can check that we are just missing 1 or 2 instances for each class. These missing values could be missing from the diffs dataset from Wikidata, or could correspond to the 2 entities that couldn't be indexed in Mongo due to their size.

### C: Number of edits per class
In this section we will try to fetch the number of operations of each class and determine if there is a correlation between their number of operations and their importance (classrank score).

The following cell returns the aggregated number of operations of the given type in a class:

def get_number_of_operations(collection, class_id, op_type=None):
    pipeline = [
        {"$match": {"class_ids": class_id}}
    ]
    
    if op_type is not None and op_type in ["add", "replace", "remove"]:
        pipeline.append({"$match": {"$op": op_type}})
    
    pipeline.append({"$unwind": "$entity_diff"})
    
    pipeline.append({
            "$group":
            {
                "_id": "$entity_diff.op",
                "count": {"$sum" : 1}
            }
    })
    return list(collection.aggregate(pipeline, allowDiskUse=True))

In the following cell we execute the previous function accross all the classes of our dataset. For each class we calculate the average number of operations by dividing the total operation count by the number of instances of the class.

> ⚠️⏰ This cell may take a long time of execute (est. 1.5h, can vary greatly based on disk speed), the precomputed result is available in the following cell and can be directly loaded without executing this cell.

data = {
    'class': [],
    'add': [],
    'remove': [],
    'replace': [],
    'classrank': []
}

with tqdm(top_classes) as t:
    for kg_class in t:
        t.set_description(f"Class: {kg_class.name}")
        num_class_ops = get_number_of_operations(wd_revisions, kg_class.qid)
        data['class'].append(kg_class.name)
        data['classrank'].append(kg_class.classrank_score)
        for op in num_class_ops:
            data[op['_id']].append(op['count'] / len(kg_class.instances))

class_ops_df = pd.DataFrame(data)
class_ops_df['total'] = class_ops_df['add'] + class_ops_df['remove'] + class_ops_df['replace']
class_ops_df.sort_values(by='total', ascending=True, inplace=True)

save_df(class_ops_df, 'c_ops')

class_ops_df = load_df('c_ops')

class_ops_df.head(n=10)

We can see from the table above the results for 10 of the classes, but in order to better visualize the dataset we are going to create some charts.

##### C.1. Number of operations per class
The following function defines the base skeleton to plot a stacked bar chart with Plotly:

def stacked_bar_chart(x_values, y, x_names, x_title, y_title, width=1000, height=1700):
    fig = go.Figure()
    for x_v, x_n in zip(x_values, x_names):
        fig.add_trace(go.Bar(
            y=y,
            x=x_v,
            name=x_n,
            orientation='h',
            hovertemplate='%{x:.2f}',
            texttemplate="%{x:.2f}"
        ))
    
    """
    totals = np.sum(x_values, axis=0)
    for total, class_name in zip(totals, y):
        fig.add_annotation(x=total, y=class_name,
                           text=f"<b>{str(round(total, 1))}<b>",
                           showarrow=False,
                           xshift=25,
                           font=dict(size=15, family="Serif"))
    """
    fig.update_layout(
        barmode='stack',
        xaxis_title=x_title,
        yaxis_title=y_title,
        legend_title="Operation type",
        height=height,
        width=width,
        font_family="Serif", font_size=24,
        uniformtext_minsize=14
    )
    return fig

We are going to use the previous function to plot the average number of operations of each class. Each stack in the bar will correspond to the operation type (addition, removal or replacement):

fig = stacked_bar_chart([class_ops_df['add'], class_ops_df['remove'], class_ops_df['replace']], class_ops_df['class'],
                        ['Additions', 'Removals', 'Replacements'], "Average operations per instance", "Class")
fig.show()

save_fig(fig, 'c_1_ops')

class_ops_df.head()

sort_class_ops_df = class_ops_df.sort_values(by=['total'], ascending=False).head(n=10).sort_values(by=['total'])
fig = stacked_bar_chart([sort_class_ops_df['add'], sort_class_ops_df['remove'], sort_class_ops_df['replace']], sort_class_ops_df['class'],
                        ['Additions', 'Removals', 'Replacements'], "Average operations per instance", "Class", height=650, width=1500)
fig.show()

save_fig(fig, 'c_1_ops_paper')

##### C.2. Correlation between operations in a class and classrank score
In this section we are going to explore a possible correlation between the number of operations of a class and its importance (classrank score). We will begin by plotting number of removal operations of each class and its classrank score, with a regression line calculated using OLS:

fig = px.scatter(class_ops_df, x="remove", y="classrank", trendline="ols", hover_name="class",
                 labels={'remove': "Average removals per instance", 'classrank': "Classrank score"})
fig.show()

save_fig(fig, 'c2_corr')

We will now check if the same trend is followed with the number of addition operations:

fig = px.scatter(class_ops_df, x="add", y="classrank", trendline="ols", hover_name="class",
                 labels={'add': "Average additions per instance", 'classrank': "Classrank score"})
fig.show()

save_fig(fig, 'c3_corr')

To finish this section we are going to calculate the pearson correlation between each variable to check if there is a correlation between the number of operations and the importance of a class.

corr_df = class_ops_df.corr(method='pearson')
corr_df

mask = np.tril(np.ones_like(corr_df, dtype=bool))

fig = go.Figure()
fig.add_trace(go.Heatmap(
    z=corr_df.mask(mask),
    x=corr_df.columns,
    y=corr_df.columns,
    colorscale=px.colors.diverging.RdBu,
    zmin=-1,
    zmax=1,
    texttemplate="%{z:.2f}")
)
fig.update_layout(width=1000, height=1000)
fig.show()

save_fig(fig, 'c3_corr_matrix')

### D: Count of properties per class
In this section we are going to check which properties are more common on each class. To do so, we are going to define the following query that aggregates the claims of each entity from a given class:

def get_properties_count(collection, class_id):
    pipeline = [
        {"$match": {"class_ids": class_id, "entity_json.claims": { "$exists": "true", "$not": {"$size": 0} }}},
        {"$project": {"claims": {"$objectToArray": "$entity_json.claims"}}},
        {"$unwind": "$claims"},
        {
            "$group":
            {
                "_id": "$claims.k", 
                "count": {"$sum" : 1}
            }
        }
    ]
    return list(collection.aggregate(pipeline, allowDiskUse=True))

In the following cell we execute the query for every class.

> ⚠️⏰ This cell may take a long time of execute (est. 2.5h, can vary greatly based on disk speed), the precomputed result is available in the following cell and can be directly loaded without executing this cell.

classes_props_data = {
    'class': [],
    'prop': [],
    'count': []
}

with tqdm(top_classes) as t:
    for kg_class in t:
        t.set_description(f"Class: {kg_class.name}")
        num_class_props = get_properties_count(wd_entities, kg_class.qid)
        for prop in num_class_props:
            classes_props_data['class'].append(kg_class.name)
            classes_props_data['prop'].append(prop['_id'])
            classes_props_data['count'].append(prop['count'] / len(kg_class.instances))
    
classes_props_df = pd.DataFrame(classes_props_data)
save_df(classes_props_df, 'd_prop_count')

classes_props_df = load_df('d_prop_count')

classes_props_df.head(n=10)

In the table above we can see the distribution of some properties of the human class.

In order to visualize this better we are going to plot the most common properties of every class. The following cell defines a function that plots a bar chart with a dropdown to select the class:

def plot_dropdown_bar_chart(x_data, y_data, labels, fig_title, x_axis_title, y_axis_title, bar_color=px.colors.qualitative.Plotly[0],
                            hovertemplate='%{x:.2f}', hovername='Count', x_range=None):
    fig = go.Figure()
    buttons = []
    i = 0
    for x, y, label in zip(x_data, y_data, labels):
        visible = i == 0 # only show first trace at start
        fig.add_trace(go.Bar(
            y=y,
            x=x,
            name=hovername,
            orientation='h',
            visible=visible,
            hovertemplate=hovertemplate,
            marker_color=bar_color
        ))

        visible_arr = [False] * len(x_data)
        visible_arr[i] = True
        i += 1
        buttons.append(dict(label=label,
                            method="update",
                            args=[{"visible": visible_arr},
                                  {"title": f"{fig_title}: {label}",
                                   "annotations": []}]))
    fig.update_layout(updatemenus=[dict(active=0, buttons=buttons, y=1.07)], title_text=f"{fig_title}: {labels[0]}", 
        xaxis_title=x_axis_title,
        yaxis_title=y_axis_title,
                     height=700)
    if x_range is not None:
        fig.update_xaxes(range=x_range)
    return fig


In the following cell we plot the data:

def plot_classes_prop_count(df, num_props):
    class_labels = df['class'].unique()
    x_data = []
    y_data = []
    for label in class_labels:
        top_props_dict = {row['prop']: row['count'] for _, row
                          in sorted(df[df['class'] == label].iterrows(), key=lambda row: row[1]['count'], reverse=True)[:num_props]
                          if row['prop'] != 'P31'}
        props = get_name_of_props(list(top_props_dict.keys()))
        count = [v * 100 for v in top_props_dict.values()]
        
        y_data.append(props[::-1])
        x_data.append(count[::-1])
    return plot_dropdown_bar_chart(x_data, y_data, class_labels, "Top properties in instances of class",
                                   "Appearance percentage in class", "Property", hovertemplate="%{x:.2f}%",
                                   hovername="Appearance", x_range=[0, 100])
    

fig = plot_classes_prop_count(classes_props_df, 25)
fig.show()

In the figure above we can inspect the most common properties from every class in the dataset.

save_fig(fig, 'd_prop_count')

### E: Properties most removed/replaced globally
In this section we are going to explore which properties are more usually edited in each class.

We will first define a set of regular expressions that can be used to match edits that modify specific parts of an entity:
- **PURE_PROPERTY_EDIT_REGEX**: This matches modifications to a complete statement group (i.e. removing all the statements of a property or adding a property for the first time).
- **PROPERTY_STATEMENTS_EDIT_REGEX**: This matches any modification to a property, but not its references or qualifiers.
- **PROPERTY_REFERENCES_EDIT_REGEX**: This matches any modification to the references of a property.
- **PROPERTY_QUALIFIERS_EDIT_REGEX**: This matches any modification to the qualifiers of a property.
- **PROPERTY_ANY_EDIT_REGEX**: This matches any type of modification to a property, including references and qualifiers.

PURE_PROPERTY_EDIT_REGEX = "\/claims\/P([0-9]*)$"
PROPERTY_STATEMENTS_EDIT_REGEX = "^(?!.*(\/hash|[0-9]\/id|\/references|\/qualifiers))\/claims\/P([0-9]*).*$"
PROPERTY_REFERENCES_EDIT_REGEX = "^(?!.*(\/hash|[0-9]\/id))\/claims\/P([0-9]*)\/[0-9]*\/references.*$"
PROPERTY_QUALIFIERS_EDIT_REGEX = "^(?!.*(\/hash|[0-9]\/id))\/claims\/P([0-9]*)\/[0-9]*\/qualifiers.*$"
PROPERTY_ANY_EDIT_REGEX = "^(?!.*(\/hash|[0-9]\/id))\/claims(\/P([0-9]*))?.*$"

The following functions will query the database to get the number of editions that match a given regular expression from above for each property both in a given class and also globally:

def get_properties_edited(collection, class_id, path_regex):
    pipeline = [
        {"$match": {"class_ids": class_id}},
        {"$unwind": "$entity_diff"},
        {"$match": {"entity_diff.path": {"$regex": path_regex}}},
        {"$addFields": {"prop": {"$regexFind": {"input": "$entity_diff.path", "regex": "P([0-9]*)"}}}},
        {
            "$group": 
            {
                "_id": {"prop": "$prop.match", "op": "$entity_diff.op"},
                "count": {"$sum" : 1}
            }
        }
    ]
    return list(collection.aggregate(pipeline, allowDiskUse=True))

def get_properties_edited_total(collection, path_regex):
    pipeline = [
        {"$unwind": "$entity_diff"},
        {"$match": {"path": {"$regex": path_regex}}},
        {"$addFields": {"prop": {"$regexFind": {"input": "$path", "regex": "P([0-9]*)"}}}},
        {
            "$group": 
            {
                "_id": {"prop": "$prop.match", "op": "$op"},
                "count": {"$sum" : 1}
            }
        }
    ]
    return list(collection.aggregate(pipeline, allowDiskUse=True))

##### E.1. Number of operations to a statement
We will start by getting the number of operations of each property for every class. As we have done previously, this count is averaged between all the instances of each class:

> ⚠️⏰ This cell may take a long time of execute (est. 5h, can vary greatly based on disk speed), the precomputed result is available in the following cell and can be directly loaded without executing this cell.

classes_claims_ops = {
    'class': [],
    'prop': [],
    'op': [],
    'count': []
}
with tqdm(top_classes) as t:
    for kg_class in t:
        t.set_description(f"Class: {kg_class.name}")
        num_class_props = get_properties_edited(wd_revisions, kg_class.qid, path_regex=PROPERTY_STATEMENTS_EDIT_REGEX)
        for prop in num_class_props:
            classes_claims_ops['class'].append(kg_class.name)
            classes_claims_ops['prop'].append(prop['_id']['prop'])
            classes_claims_ops['op'].append(prop['_id']['op'])
            classes_claims_ops['count'].append(prop['count'] / len(kg_class.instances))

classes_claims_ops_df = pd.DataFrame(classes_claims_ops)
save_df(classes_claims_ops_df, 'e1_claims_ops')

classes_claims_ops_df = load_df('e1_claims_ops')

classes_claims_ops_df.head(n=10)

##### E.2. Number of operations to a statement group
In this case we are going to get the number of operations of each statement group for every class. A statement group may contain more than one statement, and only consists of addition and removal operations (when a new 'property' is added to an entity and when all the statements of a property are removed at once from an entity):


> ⚠️⏰ This cell may take a long time of execute (est. 5h, can vary greatly based on disk speed), the precomputed result is available in the following cell and can be directly loaded without executing this cell.

classes_pure_prop_data = {
    'class': [],
    'prop': [],
    'op': [],
    'count': []
}
with tqdm(top_classes) as t:
    for kg_class in t:
        t.set_description(f"Class: {kg_class.name}")
        num_class_ops = get_properties_edited(wd_revisions, kg_class.qid, path_regex=PURE_PROPERTY_EDIT_REGEX)
        for prop in num_class_ops:
            classes_pure_prop_data['class'].append(kg_class.name)
            classes_pure_prop_data['prop'].append(prop['_id']['prop'])
            classes_pure_prop_data['op'].append(prop['_id']['op'])
            classes_pure_prop_data['count'].append(prop['count'] / len(kg_class.instances))

classes_pure_prop_df = pd.DataFrame(classes_pure_prop_data)
save_df(classes_pure_prop_df, 'e2_pure_prop_ops')

classes_pure_prop_df = load_df('e2_pure_prop_ops')

classes_pure_prop_df.head(n=10)

In the following cell we are going to plot the most created statement groups for each class. This may not be the same as the most common properties in a class, since many of these statement groups could be deleted in the final state of the entity.

def plot_most_created_properties(df, num_props):
    class_labels = df['class'].unique()
    x_data = []
    y_data = []
    for label in class_labels:
        top_props_dict = {row['prop']: row['count'] for _, row
                          in sorted(df[(df['class'] == label) & (df['op'] == 'add')].iterrows(), key=lambda row: row[1]['count'], reverse=True)[:num_props]
                          if row['prop'] != 'P31'}
        props = get_name_of_props(list(top_props_dict.keys()))
        count = [v for v in top_props_dict.values()]
        
        y_data.append(props[::-1])
        x_data.append(count[::-1])
    return plot_dropdown_bar_chart(x_data, y_data, class_labels, "Most created properties of class",
                                   "Average creations per instance", "Property")
    

fig = plot_most_created_properties(classes_pure_prop_df, 25)
fig.show()

save_fig(fig, 'e2_pure_prop_created')

Now, we are going to see which are the most deleted properties for each class:

def plot_most_deleted_properties(df, num_props):
    class_labels = df['class'].unique()
    x_data = []
    y_data = []
    for label in class_labels:
        top_props_dict = {row['prop']: row['count'] for _, row
                          in sorted(df[(df['class'] == label) & (df['op'] == 'remove')].iterrows(), key=lambda row: row[1]['count'], reverse=True)[:num_props]
                          if row['prop'] != 'P31'}
        props = get_name_of_props(list(top_props_dict.keys()))
        count = [v for v in top_props_dict.values()]
        
        y_data.append(props[::-1])
        x_data.append(count[::-1])
    return plot_dropdown_bar_chart(x_data, y_data, class_labels, "Most deleted properties of class",
                                   "Average deletions per instance", "Property", bar_color=op_colors['remove'])
    

fig = plot_most_deleted_properties(classes_pure_prop_df, 25)
fig.show()

save_fig(fig, 'e2_pure_prop_deleted')

def plot_bar_chart(df, x, y, x_title, y_title, hovername='Count', width=950, height=1100,
                   bar_color=px.colors.qualitative.Plotly[0], max_range=None, texttemplate='%{x:.2f}'):
    y = get_name_of_props(df[y])[::-1]
    y = [el if len(el) <= 26 else f"{el[:23]}..." for el in y]
    fig = go.Figure()
    fig.add_trace(go.Bar(
        y=y,
        x=list(df[x])[::-1],
        orientation='h',
        hovertemplate='%{x}',
        texttemplate=texttemplate,
        textposition='outside',
        name=hovername,
        marker_color=bar_color
    ))
    
    if max_range is None:
        max_range = max(df[x]) * 1.15

    fig.update_layout(
        xaxis_title=x_title,
        yaxis_title=y_title,
        height=height,
        width=width,
        font_family="Serif", font_size=23,
        margin=dict(l=275),
        uniformtext_minsize=20,
        xaxis=dict(range=[0, max_range], tickformat='digits')
    )
    fig.update_yaxes(automargin=False)
    return fig

def plot_most_deleted_props_of_class(df, class_name, num_props=10, height=650, width=1000):
    class_deletions_df = classes_pure_prop_df[(classes_pure_prop_df['class'] == class_name) & (classes_pure_prop_df['op'] == 'remove')].sort_values(by='count', ascending=False)
    class_deletions_df = class_deletions_df.head(n=num_props)
    return plot_bar_chart(class_deletions_df, 'count', 'prop', 'Average deletions per instance', 'Property',
                          height=height, width=width, bar_color=px.colors.qualitative.Plotly[1], max_range=2)

fig = plot_most_deleted_props_of_class(classes_pure_prop_df, 'human')
fig.show()
save_fig(fig, 'e2_pure_prop_deleted_human')

fig = plot_most_deleted_props_of_class(classes_pure_prop_df, 'taxon')
fig.show()
save_fig(fig, 'e2_pure_prop_deleted_taxon')

fig = plot_most_deleted_props_of_class(classes_pure_prop_df, 'sovereign state')
fig.show()
save_fig(fig, 'e2_pure_prop_deleted_sovereign_state')

fig = plot_most_deleted_props_of_class(classes_pure_prop_df, 'big city')
fig.show()
save_fig(fig, 'e2_pure_prop_deleted_big_city')

### F: Class operations per decile
In this section we are going to explore the dynamics with respect to time of each modification. While we have focused previously on the most created/removed properties of each class, we will now take into account the timeframe of these modifications (e.g. which properties are more created in the first moments of an entity).

To do so, we are going to define the following function that returns the modifications of an entity, grouped by the decile where they were produced:

def get_properties_edited_deciles(collection, entity_id, path_regex, decile_num=0):
    if decile_num < 0 or decile_num > 9:
        print("Decile must be a number between 0 and 9")
        return []
    
    pipeline = [
        {"$match": {"entity_id": entity_id}},
        {"$sort": {"timestamp": 1}},
        {"$bucketAuto": {
            "groupBy": "$timestamp",
            "buckets": 10,
            "output": {
                "diffs" : {"$push": "$entity_diff" }
            }
        }},
        {"$group": {
          "_id": "null",
              "buckets": {
                "$push": "$$ROOT"
              }
            }
        },
        {"$unwind": {"path": "$buckets", "includeArrayIndex": "bucketNum"}},
        {"$unwind": "$buckets.diffs"},
        {"$unwind": "$buckets.diffs"},
        {"$match": {"buckets.diffs.path": {"$regex": path_regex}}},
        {"$addFields": {"prop": {"$regexFind": {"input": "$buckets.diffs.path", "regex": "P([0-9]*)"}}}},
        {
            "$group": 
            {
                "_id": {"prop": "$prop.match", "op": "$buckets.diffs.op", "bucket": "$bucketNum"},
                "count": {"$sum" : 1}
            }
        }
    ]
    return list(collection.aggregate(pipeline, allowDiskUse=True))


In the following cell we are going to fetch this data for all the classes in our dataset.

> ⚠️⏰ This cell may take a long time of execute (est. 10h, can vary greatly based on disk speed), the precomputed result is available in the following cell and can be directly loaded without executing this cell.

from collections import defaultdict
import pdb

deciles_data = {
    'decile': [],
    'class': [],
    'prop': [],
    'op': [],
    'count': []
}

with tqdm(top_classes) as t:
    for kg_class in t:
        t.set_description(f"Class: {kg_class.name}")
        
        deciles_classes_props = {i: defaultdict(int) for i in range(10)}
        for instance in tqdm(kg_class.instances):
            num_class_props_decile = get_properties_edited_deciles(wd_revisions, instance.qid,
                                            path_regex=PROPERTY_STATEMENTS_EDIT_REGEX)

            for res in num_class_props_decile:
                key = (res['_id']['prop'], res['_id']['op'])
                deciles_classes_props[res['_id']['bucket']][key] += res['count']

        for decile in range(10):
            if decile == 0:
                continue
            
            for k, v in deciles_classes_props[decile - 1].items():
                deciles_classes_props[decile][k] += v
    
                prop, op = k[0], k[1]
                deciles_data['decile'].append(decile - 1)
                deciles_data['class'].append(kg_class.name)
                deciles_data['prop'].append(prop)
                deciles_data['op'].append(op)
                deciles_data['count'].append(v / len(kg_class.instances))
            
        for k, v in deciles_classes_props[9].items():
            prop, op = k[0], k[1]
            deciles_data['decile'].append(9)
            deciles_data['class'].append(kg_class.name)
            deciles_data['prop'].append(prop)
            deciles_data['op'].append(op)
            deciles_data['count'].append(v / len(kg_class.instances))

deciles_prop_df = pd.DataFrame(deciles_data)
save_df(deciles_prop_df, 'f_deciles_props_df')

deciles_prop_df = load_df('f_deciles_props_df')

deciles_prop_df.head(n=10)

In the following cell we are going to deploy a dashboard with all this data. This dashboard offers two endpoints: one to fetch the most edited properties of each class in a given decile, and the other to fetch this information globally:

from jupyter_dash import JupyterDash
import plotly.graph_objects as go
from dash.dependencies import Input, Output, State
from dash import callback, html, dcc
import io

n_props = 26

app = JupyterDash(__name__, suppress_callback_exceptions=True)
app.layout = html.Div([
    dcc.Location(id='url', refresh=False),
    html.Div(id='page-content')
])


index_page = html.Div([
    dcc.Link('View global data', href='/global'),
    html.Br(),
    dcc.Link('View by decile', href='/deciles'),
])

classes = deciles_prop_df['class'].unique()

page_1_layout = html.Div(
    [
        html.Div([
            "Class:",
            dcc.Dropdown(
                id="kg_class",
                options=[{"label": c, "value": c} for c in classes],
                value=classes[0],
            )
        ]),
        html.Div([
            "Operation:",
            dcc.Dropdown(
                id="op",
                options=[{"label": op, "value": op} for op in op_types],
                value=op_types[0],
            )
        ]),
        dcc.Graph(id="graph1"),
    ]
)


@app.callback(
    Output("graph1", "figure"),
    Input("kg_class", "value"),
    Input("op", "value"),
)
def update_graph1(kg_class, op):
    df = classes_claims_ops_df[classes_claims_ops_df['class'] == kg_class]

    top_props_dict = {row['prop']: row['count'] for _, row
                      in sorted(df[df['op'] == op].iterrows(), key=lambda row: row[1]['count'], reverse=True)[:n_props]
                      if row['prop'] != 'P31'}
    
    return get_figure_of(top_props_dict, op, kg_class)

page_2_layout = html.Div(
    [
        html.Div([
            "Class:",
            dcc.Dropdown(
                id="kg_class",
                options=[{"label": c, "value": c} for c in classes],
                value=classes[0],
            )
        ]),
        html.Div([
            "Decile:",
            dcc.Dropdown(
                id="decile",
                options=[{"label": i, "value": i} for i in range(1, 11)],
                value=1,
            )
        ]),
        html.Div([
            "Operation:",
            dcc.Dropdown(
                id="op",
                options=[{"label": op, "value": op} for op in op_types],
                value=op_types[0],
            )
        ]),
        dcc.Graph(id="graph2"),
    ]
)


@app.callback(
    Output("graph2", "figure"),
    Input("kg_class", "value"),
    Input("decile", "value"),
    Input("op", "value"),
)
def update_graph2(kg_class, decile, op):
    decile_idx = int(decile) - 1
    df = deciles_prop_df[deciles_prop_df['decile'] == decile_idx]
    top_props_dict = {row['prop']: row['count'] for _, row
                      in sorted(df[(df['class'] == kg_class) & (df['op'] == op)].iterrows(), key=lambda row: row[1]['count'], reverse=True)[:n_props]
                      if row['prop'] != 'P31'}
    return get_figure_of(top_props_dict, op, kg_class, decile)

def get_figure_of(props_dict, op, class_name, decile=None):
    fig = go.Figure()
    props = get_name_of_props(list(props_dict.keys()))[::-1]
    count = [v for v in props_dict.values()][::-1]

    fig.add_trace(go.Bar(
        y=props,
        x=count,
        name=f"Statements ({op})",
        orientation='h',
        hovertemplate='%{x:.2f}',
        marker_color=op_colors[op]
    ))
    
    title = f"Most edited properties for class '{class_name}'"
    if decile is not None:
        title += f" in decile {decile}"
    fig.update_layout(title_text=title, 
        xaxis_title="Average number of editions per instance",
        yaxis_title="Property",
        height=1000)
    return fig

# Update the index
@app.callback(Output('page-content', 'children'),
         [Input('url', 'pathname')])
def display_page(pathname):
    if pathname == '/global':
        return page_1_layout
    elif pathname == '/deciles':
        return page_2_layout
    else:
        return index_page

app.run_server(mode="external", host='0.0.0.0', port='8050')

## G: Conflict and general measures
In this last section we are going to fetch conflict measures and also general information of each class.

We will start by defining a set of functions to fetch all the operations of the instances of a class or of a single entity:

def get_ops_of_class(collection, class_id, path_regex, match_prop=True):
    pipeline = [
        {"$match": {"class_ids": class_id}},
        {"$project": {"entity_diff": 1, "entity_id": 1, "id": 1}},
        {"$unwind": "$entity_diff"},
        {"$match": {"entity_diff.path": {"$regex": path_regex}}}
    ]

    if match_prop:
        pipeline.append({"$addFields": {"prop": {"$regexFind": {"input": "$entity_diff.path", "regex": "P([0-9]*)"}}}})
        
    return collection.aggregate(pipeline)

def get_ops_of_entity(collection, entity_id, path_regex, match_prop=True):
    pipeline = [
        {"$match": {"entity_id": entity_id}},
        {"$project": {"entity_diff": 1, "entity_id": 1, "id": 1}},
        {"$unwind": "$entity_diff"},
        {"$match": {"entity_diff.path": {"$regex": path_regex}}}
    ]
    
    if match_prop:
        pipeline.append({"$addFields": {"prop": {"$regexFind": {"input": "$entity_diff.path", "regex": "P([0-9]*)"}}}})

    return list(collection.aggregate(pipeline))

### G.1. Fetching data
Now we will define a set of functions to detect edit wars given a set of operations. An edit war is considered a set of operations that start with a given value for a property, this value then is removed or changed to a different value, and finally the property is given the same value that it had before.

The following functions detect these edit patterns and return both the number of edit wars and the operations that compose each edit war:


def _process_edit_wars(ops, accumulator_fn):
    changes_dict = defaultdict(list)
    
    for op in ops:
        op_type = op['entity_diff']['op']
        path = op['entity_diff']['path']
        changes_dict[path].append((op_type, op['entity_diff']['value']))
    
    for path, changes in changes_dict.items():
        adds = [c for c in changes if c[0] == 'add']
        replacements = [c for c in changes if c[0] == 'replace']
        ops = adds + replacements
        
        values = []
        for op in ops:
            try:
                if isinstance(op[1], list):
                    if op[1][0]['mainsnak']['snaktype'] in ['somevalue', 'novalue']:
                        value = op[1][0]['mainsnak']['snaktype']
                    else:
                        value = op[1][0]['mainsnak']['datavalue']['value']
                elif isinstance(op[1], dict):
                    if 'mainsnak' not in op[1]:
                        value = op[1]['value']
                    elif op[1]['mainsnak']['snaktype'] in ['somevalue', 'novalue']:
                        value = op[1]['mainsnak']['snaktype']
                    else:
                        value = op[1]['mainsnak']['datavalue']['value']
                else:
                    value = op[1]
            except:
                print("Error processing op value")
                continue

            if value in values:
                accumulator_fn(path, changes)
                break
            values.append(value)

def count_edit_wars(ops):
    num_conflicts = 0
    def add_conflict():
        nonlocal num_conflicts
        num_conflicts += 1
    _process_edit_wars(ops, lambda path, changes: add_conflict())
    return num_conflicts


def detect_edit_wars(ops):
    conflicts_dict = {}
    def update_conflicts_dict(path, changes):
        nonlocal conflicts_dict
        conflicts_dict[path] = changes
    _process_edit_wars(ops, lambda path, changes: update_conflicts_dict(path, changes))
    return conflicts_dict

Now we will build two separate dataframes. The first one will contain the following information about each entity:
- ID of the entity.
- ID of the class the entity belongs to.
- Number of revisions.
- Number of operations.
- Number of 'addition' operations.
- Number of 'replacement' operations.
- Number of 'removal' operations.
- Number of edit wars detected.
- Number of statement groups added.
- Number of statement groups removed.
- Number of statements added.
- Number of statements removed.
- Number of changes in rank in one of its statements.
- Percentage of properties which are considered 'stable' (less than 5 changes to the property).
- Total number of properties.

The second one will contain the same information, but in this case about each property. The only fields not included will be the count of stable props and number of props.

all_props = classes_props_df['prop'].unique()

import pdb
import re

from collections import defaultdict

stgroups_pattern = re.compile(r"\/claims\/P([0-9]*)$")
statements_pattern = re.compile(r"\/claims\/P([0-9]*)\/([0-9]*)$")


def extract_ops_statistics(ops):
    stgroups_ops = [o for o in ops if stgroups_pattern.match(o['entity_diff']['path'])]
    statement_ops = [o for o in ops if statements_pattern.match(o['entity_diff']['path'])]
    num_edit_wars = count_edit_wars(ops)
    
    return {
        'stgroups_added': len([o for o in stgroups_ops if o['entity_diff']['op'] == 'add']),
        'stgroups_removed': len([o for o in stgroups_ops if o['entity_diff']['op'] == 'remove']),
        'statements_added': len([o for o in statement_ops if o['entity_diff']['op'] == 'add']),
        'statements_removed': len([o for o in statement_ops if o['entity_diff']['op'] == 'remove']),
        'revisions': len({o['id'] for o in ops}),
        'additions': len([o for o in ops if o['entity_diff']['op'] == 'add']),
        'replacements': len([o for o in ops if o['entity_diff']['op'] == 'replace']),
        'removals': len([o for o in ops if o['entity_diff']['op'] == 'remove']),
        'ops': len(ops),
        'edit_wars': num_edit_wars,
        'rank_changes': len([o for o in ops if '/rank' in o['entity_diff']['path']])
    }

The following cell fetches all the information of both dataframes. Everything has been joined in one loop in order to optimize the execution of the code, losing some code readability in the process.

> ⚠️⏰ This cell may take a long time of execute (est. 6h, can vary greatly based on disk speed), the precomputed result is available in the following cell and can be directly loaded without executing this cell.

revs_entities_data = {
    'entity': [], 'class': [], 'revisions': [], 'ops': [], 'edit_wars': [], 'additions': [],
    'removals': [], 'replacements': [], 'stgroups_added': [], 'stgroups_removed': [],
    'statements_added': [], 'statements_removed': [], 'rank_changes': [],
    'stable_props': [], 'all_props': []
}

revs_props_data = {
    'prop': [], 'class': [], 'revisions': [], 'ops': [], 'edit_wars': [], 'additions': [],
    'removals': [], 'replacements': [], 'statements_added': [], 'statements_removed': [],
    'stgroups_added': [], 'stgroups_removed': [], 'rank_changes': []
}

with tqdm(top_classes) as t:
    for kg_class in t:
        t.set_description(f"Class: {kg_class.name}")
        
        class_props_data = {}
        for entity in kg_class.instances:
            entity_ops = get_ops_of_entity(wd_revisions, entity.qid, path_regex=PROPERTY_STATEMENTS_EDIT_REGEX)

            if len(entity_ops) == 0:
                continue            

            revs_entities_data['entity'].append(entity.qid)
            revs_entities_data['class'].append(kg_class.qid)
            
            entity_props_ops = defaultdict(list)
            for op in entity_ops:
                entity_props_ops[op['prop']['match']].append(op)
            
            ops_statistics = defaultdict(int)
            for prop, ops in entity_props_ops.items():
                prop_statistics = extract_ops_statistics(ops)
                
                if prop not in class_props_data:
                    class_props_data[prop] = defaultdict(int)
                
                for k, v in prop_statistics.items():
                    class_props_data[prop][k] += v
                    ops_statistics[k] += v

            ops_statistics['revisions'] = len({o['id'] for o in entity_ops})
            for k, v in ops_statistics.items():
                revs_entities_data[k].append(v)
            
            # props with less than 5 'modifications' (replacements+removals)
            revs_entities_data['stable_props'].append(sum([1 for k, v in entity_props_ops.items()
                                                           if len([o for o in v if o['entity_diff']['op'] in ['replace', 'remove']]) < 5])
                                                      / len(entity_props_ops))
            revs_entities_data['all_props'].append(len(entity_props_ops))
            
        for prop, data in class_props_data.items():
            revs_props_data['class'].append(kg_class.name)
            revs_props_data['prop'].append(prop)
            for k, v in data.items():
                revs_props_data[k].append(v)

revs_entities_data_df = pd.DataFrame(revs_entities_data)
save_df(revs_entities_data_df, 'g_revs_entities_data')

revs_props_data_df = pd.DataFrame(revs_props_data)
save_df(revs_props_data_df, 'g_revs_props_data')

revs_entities_data_df = load_df('g_revs_entities_data')

revs_entities_data_df[revs_entities_data_df['entity'] == 'Q7961430']

revs_props_data_df = load_df('g_revs_props_data')

revs_entities_data_df.head(n=7)

revs_props_data_df.head(n=7)

We can see above the contents of both Dataframes.

### G.2. Analyzing conflict measures
#### Entities with the most conflictiveness
We can now see which entities have a higher conflictiveness rate from the data we have fetched. Our entities dataframe has some repeated rows (entities which belong to more than one class), so we will begin by deleting those duplicates. After that, we will calculate the mean number of edit wars per revisions of each instance. Finally, we will add a column to each row indicating the percentile that entity belong to, considering its number of edit wars:

grouped_entities_data_df = revs_entities_data_df.drop_duplicates('entity')

grouped_entities_data_df['mean_edit_wars'] = grouped_entities_data_df['edit_wars'] / grouped_entities_data_df['revisions']
grouped_entities_data_df['percentile'] = grouped_entities_data_df['edit_wars'].rank(pct=True)

fig = px.histogram(grouped_entities_data_df[grouped_entities_data_df['edit_wars'] < 200], x="edit_wars", nbins=10)
fig.show()

NUM_ELEMENTS = 15

print("Entities with the most amount of edit wars")
print("-" * 50)

entities_most_edit_wars = grouped_entities_data_df.sort_values('percentile', ascending=False).head(n=NUM_ELEMENTS)
for row in entities_most_edit_wars.itertuples():
    print(f"{get_name_of_props([row[1]])[0]}: {row.edit_wars} edit wars")

fig = plot_bar_chart(entities_most_edit_wars, 'edit_wars', 'entity', "Number of edit wars", "Entity", hovername="Edit wars",
                     width=750, height=1000, texttemplate='%{x}')
fig.show()

save_fig(fig, 'g_most_edit_wars_entities')

print("Entities with the most relative amount of edit wars")
print("-" * 50)

entities_most_edit_wars = grouped_entities_data_df.sort_values('mean_edit_wars', ascending=False).head(n=NUM_ELEMENTS)
for row in entities_most_edit_wars.itertuples():
    print(f"{get_name_of_props([row[1]])[0]}: {row.mean_edit_wars} edit wars per revision")

save_csv(grouped_entities_data_df, 'entities_data')

#### Properties with the most conflictiveness
We are now going to follow the same procedure with our properties Dataframe. In this case, we need to aggregate the values of properties accross all the classes they appear in.

grouped_revs_props_data_df = revs_props_data_df.groupby('prop', as_index=False).sum()

grouped_revs_props_data_df['mean_edit_wars'] = grouped_revs_props_data_df['edit_wars'] / grouped_revs_props_data_df['revisions']
grouped_revs_props_data_df['percentile'] = grouped_revs_props_data_df['mean_edit_wars'].rank(pct=True)

grouped_revs_props_data_df.sort_values('percentile', ascending=False).head()

print("Properties with the most amount of mean edit wars")
print("-" * 65)

props_most_edit_wars = grouped_revs_props_data_df[grouped_revs_props_data_df['additions'] > 100].sort_values('percentile', ascending=False).head(n=NUM_ELEMENTS)
for row in props_most_edit_wars.itertuples():
    print(f"{get_name_of_props([row[1]])[0]}: {round(row.mean_edit_wars, 2)} mean edit wars per revision")

fig = plot_bar_chart(props_most_edit_wars, 'mean_edit_wars', 'prop', "Number of edit wars per revision", "Property", hovername="Edit wars")
save_fig(fig, 'g_most_mean_edit_wars_properties')
fig.show()

props_most_edit_wars = grouped_revs_props_data_df.sort_values('edit_wars', ascending=False).head(n=NUM_ELEMENTS)

fig = plot_bar_chart(props_most_edit_wars, 'edit_wars', 'prop', "Number of edit wars", "Property", texttemplate='%{x}')
save_fig(fig, 'g_most_edit_wars_properties')
fig.show()

save_csv(grouped_revs_props_data_df, 'properties_data')

#### Classes with the most conflictiveness
In the following cell we are going to create a dataframe of classes by grouping all the values from the entities dataframe of each class. We are going to add a new column where the number of instances of each class is stored. In this case, the relative number of edit wars is computed based on the number of instances (i.e. mean number of edit wars per instance).

grouped_classes_data_df = revs_entities_data_df.groupby('class', as_index=False).sum()

grouped_classes_data_df['num_instances'] = grouped_classes_data_df.apply(lambda row: len([c for c in top_classes if c.qid == row['class']][0].instances), axis=1)
grouped_classes_data_df['mean_edit_wars'] = grouped_classes_data_df['edit_wars'] / grouped_classes_data_df['num_instances']

grouped_classes_data_df['percentile'] = grouped_classes_data_df['mean_edit_wars'].rank(pct=True)

grouped_classes_data_df.sort_values('percentile', ascending=False).head()

print("Classes with the most amount of edit wars")
print("-" * 65)

classes_most_edit_wars = grouped_classes_data_df.sort_values('percentile', ascending=False).head(n=NUM_ELEMENTS)
for row in classes_most_edit_wars.itertuples():
    print(f"{get_name_of_props([row[1]])[0]}: {round(row.mean_edit_wars, 2)} mean edit wars per instance")

fig = plot_bar_chart(classes_most_edit_wars, 'mean_edit_wars', 'class', "Mean number of edit wars per instance", "Class", hovername="Edit wars")
save_fig(fig, 'g_most_edit_wars_classes')
fig.show()

print("Classes with the least amount of edit wars")
print("-" * 65)

classes_least_edit_wars = grouped_classes_data_df.sort_values('percentile', ascending=True).head(n=NUM_ELEMENTS)
for row in classes_least_edit_wars.itertuples():
    print(f"{get_name_of_props([row[1]])[0]}: {round(row.mean_edit_wars, 3)} mean edit wars per instance")

save_csv(grouped_classes_data_df, 'classes_data')

### Correlation between variables
Finally, we will calculate the correlation between the variables of our entities dataframe using the Pearson method:

corr_revs_entities_data_df = revs_entities_data_df.corr(method='pearson')

mask = np.tril(np.ones_like(corr_revs_entities_data_df, dtype=bool))

fig = go.Figure()
fig.add_trace(go.Heatmap(
    z=corr_revs_entities_data_df.mask(mask),
    x=corr_revs_entities_data_df.columns,
    y=corr_revs_entities_data_df.columns,
    colorscale=px.colors.diverging.RdBu,
    zmin=-1,
    zmax=1,
    texttemplate="%{z:.2f}")
)
fig.update_layout(width=1000, height=1000)
fig.show()

### H: Life cycle of an entity
In the following cells we are going to plot which are the most frequent operations made to entities in our dataset, as well as the most common transitions between these operations.

We will start by defining the code that categorizes each operation and computes the transition counts form one operation to another:

import collections
import re

path_to_state = collections.OrderedDict({
    re.compile(r"^\/claims\/P([0-9]*)$"): "statement group",
    re.compile(r"^\/claims\/P([0-9]*)\/([0-9]*)$"): "statement",
    re.compile(r"^\/claims\/P([0-9]*)\/([0-9]*)\/mainsnak((\/datavalue)|(\/snaktype)).*$"): "statement value",
    re.compile(r"^\/claims\/P([0-9]*)\/([0-9]*)\/rank$"): "statement rank",
    re.compile(r"^\/claims\/P([0-9]*)\/([0-9]*)\/references(\/([0-9]*))?$"): "reference",
    re.compile(r"^\/claims\/P([0-9]*)\/([0-9]*)\/references\/([0-9]*)\/snaks\/P([0-9]*)(\/([0-9]*))?$"): "reference snak",
    re.compile(r"^\/claims\/P([0-9]*)\/([0-9]*)\/references\/([0-9]*)\/snaks-order\/.*$"): "reference order",
    re.compile(r"^\/claims\/P([0-9]*)\/([0-9]*)\/references\/([0-9]*)\/snaks\/P([0-9]*)\/([0-9]*)((\/datavalue)|(\/snaktype)).*$"): "reference value",
    re.compile(r"^\/claims\/P([0-9]*)\/([0-9]*)\/qualifiers$"): "qualifier",
    re.compile(r"^\/claims\/P([0-9]*)\/([0-9]*)\/qualifiers\/(P[0-9]*)(\/([0-9]*))?$"): "qualifier snak",
    re.compile(r"^\/claims\/P([0-9]*)\/([0-9]*)\/qualifiers-order(\/([0-9]*))?$"): "qualifier order",
    re.compile(r"^\/claims\/P([0-9]*)\/([0-9]*)\/qualifiers\/P([0-9]*)\/([0-9]*)((\/datavalue)|(\/snaktype)).*$"): "qualifier value",
    re.compile(r"^\/claims$"): "entity",
})

states = [(s, t) for s in list(path_to_state.values()) for t in ['add', 'replace', 'remove']]
states_idx = {}
for i, s in enumerate(states):
    states_idx[s] = i

def get_operation_state(op):
    path_type = "other"
    for k in path_to_state:
        if k.match(op['path']):
            path_type = path_to_state[k]
            break
    
    op_type = op['op']
    if '/snaktype' in op['path']:
        op_type = 'remove' if op['value'] == 'novalue' else 'add'
        
    
    res = (path_type, op_type)

    if path_type == "other":
        print(f"Input path: {op['path']}")
        print(f"Output type: {res}")

    return (path_type, op_type)


def build_state_count(ops):
    state_counts = np.zeros((len(states), len(states)))
    prev_state = None
    
    for op in ops:
        curr_state = states_idx[get_operation_state(op['entity_diff'])]
        if prev_state is not None:
            state_counts[prev_state, curr_state] += 1
            
        prev_state = curr_state
    return state_counts

In the following cell we query the database to fetch this information:

> ⚠️⏰ This cell may take a long time of execute (est. 7h, can vary greatly based on disk speed), the precomputed result is available in the following cell and can be directly loaded without executing this cell.

import pdb

state_counts = np.zeros((len(states), len(states)))
with tqdm(top_classes) as t:
    for kg_class in t:
        t.set_description(f"Class: {kg_class.name}")
        for entity in tqdm(kg_class.instances[:3000000]):
            entity_ops = get_ops_of_entity(wd_revisions, entity.qid, path_regex=PROPERTY_ANY_EDIT_REGEX, match_prop=False)
            state_counts = np.add(state_counts, build_state_count(entity_ops))

save_df(state_counts, 'h_state_count')
state_counts

state_counts = load_df('h_state_count')
state_counts

remove_list = [18, 20, 30, 32, 35]
for el in remove_list:
    state_name = {v: k for k, v in states_idx.items()}[el]
    del states_idx[state_name]
state_counts = np.delete(state_counts, remove_list, 0)
state_counts = np.delete(state_counts, remove_list, 1)


i = 0
for s in states:
    if s in states_idx:
        states_idx[s] = i
        i += 1

#### Graph visualization
Finally, this cell uses the pyvis library to plot the state counts that were computed before:

from pyvis import network as net
import math

max_num = np.max(state_counts)
min_num = np.min(state_counts[state_counts > 0])

max_incoming = max(np.sum(state_counts, axis=0))
min_incoming = min(np.sum(state_counts, axis=0))


g = net.Network(notebook=True, directed=True, height='750px', width='750px')

# add nodes
for s, i in states_idx.items():
    if s == ('entity', 'remove') or s == ('qualifier value', 'remove'):
        # we are not recording entity removals, since they are not available in the revision history data
        continue
    
    if sum(state_counts[i, :]) + sum(state_counts[:, i]) > 0:
        if s == ('entity', 'add'):
            color = "#8b5cf6"
            label = "Create entity"
        else:
            color = "#60a5fa"
            label = f"Modify {s[0]}" if 'order' in s[0] or 'rank' in s[0] else f"{s[1].capitalize()} {s[0]}"
        incoming_count = sum(state_counts[:, i])
        if s == ('entity', 'add'):
            # add all entities that have been created to count
            incoming_count += 9300000
        theta = 10
        node_scale = math.log(((min_incoming * theta - max_incoming) / (min_incoming - max_incoming)) + (((1-theta) * incoming_count)/ (min_incoming - max_incoming)), theta)

        g.add_node(i, label=label, color=color, borderWidth=1.5, size=max(15, node_scale * 65))

for s1, i1 in states_idx.items():
    if s1 == ('entity', 'remove') or s1 == ('qualifier value', 'remove'):
        continue
    
    all_edges_s1 = sum(state_counts[i1, :])    
    for s2, i2 in states_idx.items():
        weight = state_counts[i1, i2]

        if weight <= 0:
            continue
        
        rel_weight = weight / all_edges_s1
        theta = 100
        graph_weight = math.log(((min_num * theta - max_num) / (min_num - max_num)) + (((1-theta) * weight)/ (min_num - max_num)), theta)
        
        if rel_weight > 0.11:
            g.add_edge(i1, i2, width=max(0.25, graph_weight * 4), title=graph_weight)

for s1, i1 in states_idx.items():
    if s1 == ('entity', 'add') or s1 == ('entity', 'remove') or s1 == ('qualifier value', 'remove'):
        continue
    
    all_edges_s1 = sum(state_counts[:, i1])
    for s2, i2 in states_idx.items():
        weight = state_counts[i2, i1]
        if weight <= 0:
            continue
        
        rel_weight = weight / all_edges_s1
        if rel_weight > 0.11:
            theta = 100
            graph_weight = math.log(((min_num * theta - max_num) / (min_num - max_num)) + (((1-theta) * weight)/ (min_num - max_num)), theta)
            all_edges = g.get_edges()
            has_edge = False
            for e in all_edges:
                if e['from'] == i2 and e['to'] == i1:
                    has_edge = True
            
            if not has_edge:
                g.add_edge(i2, i1, width=max(0.25, graph_weight * 4), title=graph_weight)

g.set_options("""
var options = {
  "nodes": {
    "font": {
      "face": "verdana",
      "strokeWidth": 3,
      "size": 24
    }
  },
  "edges": {
    "color": {
      "inherit": true
    },
    "font": {
      "face": "verdana"
    },
    "smooth": {
      "type": "continuous",
      "forceDirection": "none",
      "roundness": 0.25
    }
  },
  "physics": {
    "forceAtlas2Based": {
      "gravitationalConstant": -277,
      "springLength": 150,
      "springConstant": 0.06
    },
    "minVelocity": 0.76,
    "solver": "forceAtlas2Based"
  }
}
""")

g.show(os.path.join(OUTPUT_DIR, "edit_states_graph.html"))
